+++ apim-server/server/routes.ts
@@ -9,15 +9,21 @@ import verifyServerStatus from './api/middlewares/verifyServerStatus';
 
 export default function routes(app: Application, apiBase: string): void {
   const router = Router();
-
+  // unprotected routes
   // available even if server not operational
   router.use('/apsMonitor', apsMonitorRouter);
   // check that server is ready
   router.use(verifyServerStatus);
+  // protected routes
   router.use('/apsUsers', apsUsersRouter);
   router.use('/apsLogin', apsLoginRouter);
   router.use('/apsConfig/apsConnectors', apsConnectorRouter);
   router.use('/apsConfig/apsAbout', apsAboutRouter);
+  // TODO: Test passport
+  // const passport = PassportFactory.build();
+  // router.use(passport.initialize());
+  // router.use('/*', passport.authenticate(['provider', 'basic'], PassportFactory.getAuthenticationOptions()));
+  // router.use('/apsApps')
 
   app.use(apiBase, router);
 }